requirements.txt
pytest==8.1.1
pytest-cov
requests
pytest>=7.0
pytest-cov>=4.0
colorama>=0.4.6

fastapi
uvicorn[standard]
sqlalchemy
alembic
python-dotenv
pydantic[email]>=2.0
pydantic-settings>=2.0
mysql-connector-python
httpx
email-validator
passlib
python-jose[cryptography]
pymysql